add lwt.unix to src/core build to support lwt>=4
This is a separate module in Lwt>=4.0, so needs to be explicitly
mentioned in the jbuild file as it is no longer in the same directory as
Lwt itself.

Note that this is probably a bug in github-hooks, since the core library
is not meant to be depend on Unix, but that fix is a bigger one since
it requires propagating a timeout function through the module.  This
PR just fixes the build for now.